+++ pe-bear/gui/PeWrapperModel.h
@@ -32,7 +32,7 @@ public:
 	virtual bool setData(const QModelIndex &index, const QVariant &value, int role);
 	virtual Qt::ItemFlags flags(const QModelIndex &index) const;
 
-	virtual QString makeDockerTitle(uint32_t upId);
+	virtual QString makeDockerTitle(size_t upId);
 	
 	virtual int getFID(const QModelIndex &index) const
 	{
