+++ pe-bear/gui_base/WrapperTableModel.cpp
@@ -57,7 +57,7 @@ Qt::ItemFlags WrapperTableModel::flags(const QModelIndex &index) const
 	return f;
 }
 
-QString WrapperTableModel::makeDockerTitle(uint32_t upId)
+QString WrapperTableModel::makeDockerTitle(size_t upId)
 {
 	ExeNodeWrapper* node = dynamic_cast<ExeNodeWrapper*>(wrapper());
 	if (node == NULL) {
@@ -68,7 +68,7 @@ QString WrapperTableModel::makeDockerTitle(uint32_t upId)
 		return "-";
 	}
 	QString name = childEntry->getName();
-	uint32_t funcNum = childEntry->getEntriesCount();
+	size_t funcNum = childEntry->getEntriesCount();
 	QString numDesc = funcNum == 1 ? tr(" entry") : tr(" entries");
 	QString desc = name + "   [ " + QString::number(funcNum) + numDesc + " ]"; 
 	return desc;
